Add mpasa60 domain and mapping files to config files
This commit adds the mpasa60 domain and mapping files to the CIME config files.
It should be noted that mapping files were made with the tx0.1v2 mask, which is
different than the other MPAS grids. This may be incorrect.

This commit does not add a NUOPC grid to the configuration files.